feat: add manage dialog in present move
