+++ src/renderer/components/media/MediaControls.vue
@@ -1,6 +1,19 @@
 <!-- Media controls for the presentation mode -->
 <template>
   <v-row>
+    <v-dialog :value="manageMedia" fullscreen>
+      <v-sheet :color="isDark ? '#121212' : '#ffffff'" class="fill-height">
+        <v-container fluid fill-height>
+          <manage-media
+            :media="localMedia"
+            :loading="loading"
+            dialog
+            @cancel="manageMedia = false"
+            @refresh="getMedia()"
+          />
+        </v-container>
+      </v-sheet>
+    </v-dialog>
     <present-top-bar
       :media-active="mediaActive"
       :current-index="currentIndex"
@@ -14,6 +27,7 @@
       @sortable="sortable = !sortable"
       @prefix="togglePrefix()"
       @refresh="getMedia()"
+      @manage-media="manageMedia = true"
     />
     <loading-icon v-if="loading" />
     <media-list
@@ -38,6 +52,7 @@ import { defineComponent } from 'vue'
 import { basename, dirname, join } from 'upath'
 import { ipcRenderer } from 'electron'
 import { MS_IN_SEC } from '~/constants/general'
+import { LocalFile } from '~/types'
 type MediaItem = {
   id: string
   path: string
@@ -66,6 +81,7 @@ export default defineComponent({
   },
   data() {
     return {
+      manageMedia: false,
       currentIndex: -1,
       sortable: false,
       loading: true,
@@ -81,6 +97,18 @@ export default defineComponent({
     },
     mediaVisible(): boolean {
       return this.$store.state.present.mediaScreenVisible
+    },
+    localMedia(): LocalFile[] {
+      return this.items.map((item) => {
+        return {
+          safeName: basename(item.path),
+          isLocal: true,
+          filepath: item.path,
+        }
+      })
+    },
+    isDark(): boolean {
+      return this.$vuetify.theme.dark as boolean
     },
     scene(): string {
       return this.$store.state.obs.currentScene as string
